Revert "Merge branch 'oscar-system:master' into Scheme_development"
This reverts commit 01ada6362220a1578d11ef0331f4b940f36600c6, reversing
changes made to 79007d39202cff1cd3b6c7503cbe3e89cf732ba1.
