+++ .github/workflows/CI.yml
@@ -16,7 +16,6 @@ jobs:
           - '1.4'
           - '1.5'
           - '1.6'
-          - '~1.7.0-0'
           - 'nightly'
         julia-arch:
           - x64
