+++ Project.toml
@@ -19,9 +19,9 @@ Singular = "bcd08a7b-43d2-5ff7-b6d4-c458787f915c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
-AbstractAlgebra = "0.20.0"
+AbstractAlgebra = "0.19.0"
 DocStringExtensions = "0.8"
-GAP = "0.6"
+GAP = "0.5"
 Hecke = "0.10.12"
 Nemo = "0.25.0"
 Polymake = "0.5.6"
