--- docs/src/NumberTheory/intro.md
@@ -1,18 +1,0 @@
-```@meta
-CurrentModule = Oscar
-```
-
-```@setup oscar
-using Oscar
-```
-
-```@contents
-Pages = ["intro.md"]
-```
-
-# Introduction
-
-The number theory part of OSCAR provides functionality for algebraic number theory.
-
-It is under development with regard to providing both the functionality and the documentation. 
-
