+++ experimental/GaloisGrp/GaloisGrp.jl
@@ -12,8 +12,7 @@ import Hecke: orbit, fixed_field
 
 
 function __init__()
-  GAP.Packages.install("ferret"; interactive=false, quiet=true)
-  GAP.Packages.load("ferret")
+  GAP.Packages.load("ferret", install = true)
 
   Hecke.add_verbose_scope(:GaloisGroup)
   Hecke.add_verbose_scope(:GaloisInvariant)
