+++ experimental/PlaneCurve/ProjPlaneCurve.jl
@@ -16,32 +16,6 @@ export issmooth, tangent, common_components, curve_intersect,
     issmooth(C::ProjectivePlaneCurve{S}, P::Oscar.Geometry.ProjSpcElem{S}) where S <: FieldElem
 
 Throw an error if `P` is not a point of `C`, return `false` if `P` is a singular point of `C`, and `true` if `P` is a smooth point of `C`.
-
-# Example
-```jldoctest
-julia> S, (x, y, z) = PolynomialRing(QQ, ["x", "y", "z"])
-(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
-
-julia> T, _ = grade(S)
-(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
-  x -> [1]
-  y -> [1]
-  z -> [1], MPolyElem_dec{fmpq,fmpq_mpoly}[x, y, z])
-
-julia> C = Oscar.ProjPlaneCurve(x^2*(x+y)*(y^3-x^2*z))
-Projective plane curve defined by -x^5*z - x^4*y*z + x^3*y^3 + x^2*y^4
-
-
-julia> PP = projective_space(QQ, 2)
-(Projective space of dim 2 over Rational Field
-, MPolyElem_dec{fmpq,fmpq_mpoly}[x[0], x[1], x[2]])
-
-julia> P = Oscar.Geometry.ProjSpcElem(PP[1], [QQ(0), QQ(0), QQ(1)])
-(0 : 0 : 1)
-
-julia> Oscar.issmooth(C, P)
-false
-```
 """
 function Oscar.issmooth(C::ProjectivePlaneCurve{S}, P::Oscar.Geometry.ProjSpcElem{S}) where S <: FieldElem
   dim(P.parent) == 2 || error("The point needs to be in a projective two dimensional space")
@@ -62,32 +36,6 @@ end
     tangent(C::ProjectivePlaneCurve{S}, P::Oscar.Geometry.ProjSpcElem{S}) where S <: FieldElem
 
 Return the tangent of `C` at `P` when `P` is a smooth point of `C`, and throw an error otherwise.
-
-# Example
-```jldoctest
-julia> S, (x, y, z) = PolynomialRing(QQ, ["x", "y","z"])
-(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
-
-julia> T, _ = grade(S)
-(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
-  x -> [1]
-  y -> [1]
-  z -> [1], MPolyElem_dec{fmpq,fmpq_mpoly}[x, y, z])
-
-julia> PP = projective_space(QQ, 2)
-(Projective space of dim 2 over Rational Field
-, MPolyElem_dec{fmpq,fmpq_mpoly}[x[0], x[1], x[2]])
-
-julia> C = Oscar.ProjPlaneCurve(x^2*(x+y)*(y^3-x^2*z))
-Projective plane curve defined by -x^5*z - x^4*y*z + x^3*y^3 + x^2*y^4
-
-
-julia> P = Oscar.Geometry.ProjSpcElem(PP[1], [QQ(2), QQ(-2), QQ(1)])
-(2 : -2 : 1)
-
-julia> Oscar.tangent(C, P)
-Projective plane curve defined by -48*x - 48*y
-```
 """
 function tangent(C::ProjectivePlaneCurve{S}, P::Oscar.Geometry.ProjSpcElem{S}) where S <: FieldElem
   dim(P.parent) == 2 || error("The point needs to be in a projective two dimensional space")
@@ -144,35 +92,6 @@ end
     curve_intersect([PP::Oscar.Geometry.ProjSpc{S}], C::ProjectivePlaneCurve{S}, D::ProjectivePlaneCurve{S}) where S <: FieldElem
 
 Return a list whose first element is the projective plane curve defined by the gcd of `C.eq` and `D.eq`, the second element is the list of the remaining intersection points when the common components are removed from `C` and `D` (the points are in `PP` if specified, or in a new projective space otherwise).
-
-# Example
-```jldoctest
-julia> S, (x, y, z) = PolynomialRing(QQ, ["x", "y","z"])
-(Multivariate Polynomial Ring in x, y, z over Rational Field, fmpq_mpoly[x, y, z])
-
-julia> T, _ = grade(S)
-(Multivariate Polynomial Ring in x, y, z over Rational Field graded by
-  x -> [1]
-  y -> [1]
-  z -> [1], MPolyElem_dec{fmpq,fmpq_mpoly}[x, y, z])
-
-julia> PP = projective_space(QQ, 2)
-(Projective space of dim 2 over Rational Field
-, MPolyElem_dec{fmpq,fmpq_mpoly}[x[0], x[1], x[2]])
-
-julia> C = Oscar.ProjPlaneCurve(T(x+y+z))
-Projective plane curve defined by x + y + z
-
-
-julia> D = Oscar.ProjPlaneCurve(T(z))
-Projective plane curve defined by z
-
-
-julia> Oscar.curve_intersect(PP[1], C, D)
-2-element Array{Array{Any,1},1}:
- []
- [(-1 : 1 : 0)]
-```
 """
 function curve_intersect(PP::Oscar.Geometry.ProjSpc{S}, C::ProjectivePlaneCurve{S}, D::ProjectivePlaneCurve{S}) where S <: FieldElem
   G = gcd(C.eq, D.eq)
@@ -459,13 +378,13 @@ function arithmetic_genus(C::ProjectivePlaneCurve)
   I = ideal(S, [F])
   A = quo(S, I)
   H = hilbert_polynomial(A[1])
-  return -ZZ(coeff(H, 0)) + 1
+  return -coeff(H, 0) + 1
 end
 
 ################################################################################
 
 @doc Markdown.doc"""
-    geometric_genus(C::ProjectivePlaneCurve{S}) where S <: FieldElem
+    geometric_genus(C::ProjectivePlaneCurve)
 
 Return the geometric genus of `C`.
 """
@@ -475,12 +394,12 @@ function geometric_genus(C::ProjectivePlaneCurve{S}) where S <: FieldElem
    I = ideal(R, [F])
    if S == fmpq
       singular_assure(I)
-      return ZZ(Singular.LibNormal.genus(I.gens.S)::Int)
+      return Singular.LibNormal.genus(I.gens.S)
    else
       A = quo(R, I)
       L = normalization(A[1])
       m = length(L)
-      pa = zero(ZZ)
+      pa = 0
       for i in 1:m
          J = L[i][1].I
          T, _ = grade(parent(J[1]))
@@ -488,7 +407,7 @@ function geometric_genus(C::ProjectivePlaneCurve{S}) where S <: FieldElem
          JJ = ideal(T, V)
          B = quo(T, JJ)
          H = hilbert_polynomial(B[1])
-         pa = pa - ZZ(coeff(H, 0))
+         pa = pa - coeff(H, 0)
       end
       return pa + 1
    end
