--- src/InvariantTheory/InvariantTheory.jl
@@ -1,1 +1,0 @@
-include("invariant_rings.jl")
