+++ test/runtests.jl
@@ -22,10 +22,9 @@ include("Groups/runtests.jl")
 include("Geometry/Schemes.jl")
 include("Rings/NumberField.jl")
 include("Rings/FunctionField-test.jl")
-include("Rings/AbelianClosure.jl")
 
 if Oscar.is_dev
-  include("Examples/PlaneCurve-test.jl")
+   include("Examples/PlaneCurve-test.jl")
 end
 
 include("Rings/binomial-ideals-test.jl")
